finishing Weight in Feature, correction type Double and call function

+++ src/main/scala/edu/gatech/cse8803/model/models.scala
@@ -16,7 +16,7 @@ case class LabResult(patientID: String, date: Long, labName: String, loincCode: 
 case class Diagnostic(patientID: String, date: Long, icd9code: String) //, sequence: Int
 
 case class Medication(patientID: String, date: Long, medicine: String)
-case class Vital(patientID: String, date: Long, Height: Int, Weight: Int, SystolicBP: Int, DiastolicBP: Int, Pulse: Int, Respiration: Int, Temperature: Int)
+case class Vital(patientID: String, date: Long, Height: Double, Weight: Double, SystolicBP: Double, DiastolicBP: Double, Pulse: Double, Respiration: Double, Temperature: Double)
 
 case class DataSet(patientID: String, label: String, featureVector: Vector)
